Improve the performance of the 'tf-idf' calculations

+++ src/main/kotlin/tfidf/TermFrequencyCalculator.kt
@@ -6,12 +6,12 @@ import codes.jakob.semanticcoupling.model.Term
 
 class TermFrequencyCalculator(private val document: Document) {
     fun calculate(termToCheck: Term): Double {
-        var frequency = 0.0
+        var frequency = 0
 
         for (term: Term in document.terms) {
             if (termToCheck == term) frequency++
         }
 
-        return (frequency / document.terms.size)
+        return (frequency.toDouble() / document.terms.size)
     }
 }
